ferramas-system/usuario_service/requirements.txt
whitenoise==6.5.0
anyio==4.9.0
asgiref==3.8.1
click==8.2.0
Django==5.2.1
djangorestframework==3.16.0
h11==0.16.0
httptools==0.6.4
idna==3.10
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
django-cors-headers>=4.7.0
mysqlclient>=2.1.0
drf-yasg>=1.20.0

